Refactoring & Rearranging for Production-Readiness
