+++ src/app/api/allusions/allusionQuery/route.dev.js
@@ -1,5 +1,5 @@
-const { getSession } = require('../../neo4j_driver/route.dev.js');
-const { toNativeTypes, valueToNativeType,concatObj, sortPnumsFromObjList } = require('../../neo4j_driver/utils.dev.js');
+const { getSession } = require('../../neo4j_driver/route.prod.js');
+const { toNativeTypes, valueToNativeType,concatObj, sortPnumsFromObjList } = require('../../neo4j_driver/utils.prod.js');
 
 
 async function getData (){
