+++ src/app/api/allusions/route.dev.js
@@ -1,2 +1,2 @@
-const { getSession } = require('../neo4j_driver/route.dev.js');
-const { toNativeTypes,concatObj,getChpList, sortPnumsFromObjLis } = require('../neo4j_driver/utils.dev.js');
+const { getSession } = require('../neo4j_driver/route.prod.js');
+const { toNativeTypes,concatObj,getChpList, sortPnumsFromObjLis } = require('../neo4j_driver/utils.prod.js');
