+++ src/app/api/search/search_table/route.prod.js
@@ -1,6 +1,6 @@
-const { getSession, getChpList } = require('../neo4j_driver/route.dev.js');
+const { getSession, getChpList } = require('../../neo4j_driver/route.prod.js');
 
-import { toNativeTypes } from '../neo4j_driver/utils.dev.js';
+import { toNativeTypes } from '../../neo4j_driver/utils.prod.js';
 
 async function getData (){
     const session = await getSession()
