+++ src/app/poems/[chapter]/[poem]/page.prod.jsx
@@ -1,6 +1,6 @@
 'use client'
 
-import PoemQueryResults from '@/components/PoemQueryResults.dev'
+import PoemQueryResults from '@/components/PoemQueryResults.prod'
 
 import styles from "../../../../styles/pages/poems.module.css"
